add auth & paginate

+++ model/index.js
@@ -4,6 +4,7 @@
 // const { v4: uuid } = require("uuid");
 
 const { ResumeToken } = require("mongodb");
+const { getMaxListeners } = require("../app");
 const Contact = require("./schemas/contact");
 
 // const getCollection = async (db, name) => {
@@ -12,17 +13,45 @@ const Contact = require("./schemas/contact");
 //   return collection;
 // };
 
-const listContacts = async () => {
-  const results = await Contact.find({});
-  return results;
+const listContacts = async (userId, query) => {
+  const {
+    limit = 5,
+    offset = 0,
+    sortBy,
+    sortByDesc,
+    filter, // name|favorite|email
+    favorite = null,
+  } = query;
+  const optionSearch = { owner: userId };
+  if (favorite !== null) {
+    optionSearch.favorite = favorite;
+  }
+  const results = await Contact.paginate(optionSearch, {
+    limit, // Лимит выдачи
+    offset, // Смещение на кол-во
+    select: filter ? filter.split("|").join(" ") : "", // Выбор отображаемых полей
+    sort: {
+      ...(sortBy ? { [`${sortBy}`]: 1 } : {}), // Сортировака по возрастанию
+      ...(sortByDesc ? { [`${sortByDesc}`]: -1 } : {}), // Сортировака по убывани
+    },
+  });
+
+  // Пример запроса
+  // http://localhost:3000/api/contacts?limit=5&favorite=false&filter=name|favorite|email&sortByDesc=name
+
+  const { docs: contacts, totalDocs: total } = results;
+  return { contacts, total, limit, offset };
   // const collection = await getCollection(db, "contacts");
   // const results = collection.find({}).toArray();
   // return results;
   // return db.get("contacts").value();
 };
 
-const getContactById = async (id) => {
-  const results = await Contact.findOne({ _id: id });
+const getContactById = async (userId, id) => {
+  const results = await Contact.findOne({ _id: id, owner: userId }).populate({
+    path: "owner",
+    select: "name email subscription -_id",
+  });
   return results;
   // const collection = await getCollection(db, "contacts");
   // const [results] = await collection.find({ _id: new ObjectId(id) }).toArray();
@@ -30,8 +59,8 @@ const getContactById = async (id) => {
   // return db.get("contacts").find({ id }).value();
 };
 
-const removeContact = async (id) => {
-  const result = await Contact.findByIdAndRemove({ _id: id });
+const removeContact = async (userId, id) => {
+  const result = await Contact.findByIdAndRemove({ _id: id, owner: userId });
   // const collection = await getCollection(db, "contacts");
   // const { value: result } = await collection.findOneAndDelete({
   //   _id: new ObjectId(id),
@@ -55,11 +84,14 @@ const addContact = async (body) => {
   return result;
 };
 
-const patchContact = async (id, body) => {
+const patchContact = async (userId, id, body) => {
   const result = await Contact.findOneAndUpdate(
-    { _id: id },
+    {
+      _id: id,
+      owner: userId,
+    },
     { ...body },
-    { new: false }
+    { new: true }
   );
 
   // const collection = await getCollection(db, "contacts");
@@ -80,7 +112,7 @@ const patchContact = async (id, body) => {
   return result;
 };
 
-const updateContact = async (id, body, next) => {
+const updateContact = async (userId, id, body, next) => {
   console.log(body);
   if (Object.keys(body).length === 0) {
     next({
@@ -90,9 +122,12 @@ const updateContact = async (id, body, next) => {
   }
 
   const result = await Contact.findOneAndUpdate(
-    { _id: id },
+    {
+      _id: id,
+      owner: userId,
+    },
     { ...body },
-    { new: false }
+    { new: true }
   );
 
   // const collection = await getCollection(db, "contacts");
